Core change to enrol_database: clone courses templated from {local_xmlsync_crsimport} records.
Creates:
    \local_xmlsync\util::enrol_database_template_check()
    \local_xmlsync\util::enrol_database_template_hook()
    \local_xmlsync\util::enrol_database_course_hook()
    \local_xmlsync\util::enrol_database_course_update_hook()

If the course_idnumber matches and the course_template field is a
valid course, create a copy.

Copies are not immediate; they are triggered via the sync_courses()
method and will be completed on a cron run.

This also checks visibility of the course from the import system
and sets/updates it

WR#371793